Put textarea in a wrapper component
